+++ breeze-data/src/main/kotlin/com/windea/breezeframework/data/domain/TEntity.kt
@@ -3,10 +3,9 @@ package com.windea.breezeframework.data.domain
 import com.windea.breezeframework.core.extensions.*
 import com.windea.breezeframework.data.extensions.*
 import java.io.*
-import java.util.*
 
 /**泛型实体类。此类的子类应当是开放的。*/
-@Deprecated("使用`list.distinct{ it.id }`保证唯一性，使用`entity.toPropertyMap()`得到属性信息。")
+@Deprecated("使用`equalsByDistinct()`、`hashcodeBySelect()`和`toStringBySelect()`。")
 abstract class TEntity<ID> : Serializable {
 	open var id: ID? = null
 	
@@ -17,7 +16,7 @@ abstract class TEntity<ID> : Serializable {
 	}
 	
 	override fun hashCode(): Int {
-		return Objects.hash(super.hashCode(), this.id.hashCode())
+		return this.id.hashCode()
 	}
 	
 	override fun toString(): String {